Make FakeInfoDictionaryObjectProvider reusable
